backend-deploy/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Pillow==9.5.0
requests==2.31.0
python-multipart==0.0.6